requirements.txt
torch 
transformers 
llama-cpp-python 
numpy 
pandas 
matplotlib 
seaborn 
wordcloud 
scikit-learn 
